refactor(adapters,hooks,presence): deduplicate shared utilities after review
- Extract createSoftUnmountScheduler to adapter-base and use it in React/Vue adapters
- Extract createInstanceTreeMarkers to adapter-base and use it in React/Vue/WC adapters
- Hoist validTransitions to module scope in as-transition to avoid recreating it per call
- Add isPromiseLike helper in presence module and replace 5 repeated duck-typing checks
