+++ packages/adapters/base/src/index.ts
@@ -4,6 +4,8 @@ export * from './wiring/caps-builder';
 export * from './wiring/dom-order-observer';
 export * from './gate/event-gate';
 export * from './lifecycle/teardown';
+export * from './lifecycle/soft-unmount';
 export * from './host/adapter-host';
 export * from './events/web-event-router';
+export * from './platform/instance-tree';
 export * from './types';
